requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
langgraph==0.2.40
langchain-openai==0.2.8
langchain-core==0.3.20
requests==2.32.3
pydantic==2.9.2
python-multipart>=0.0.6